Only use 2 sections for imports
Many people (myself included) prefer to have a third section to the
imports for all project-local packages. This is achievable with flags
such as `-local` on goimports. However experience shows that developers
do not always have their editors configured to do this so there tends to
be "git noise" as some developers enforce this pattern and others don't.

The 2 section imports seems to be the lowest common denominator.